requirements.txt
asgiref==3.6.0
bleach==6.0.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.1.0
contourpy==1.2.0
cryptography==41.0.3
cycler==0.12.1
dateparser==1.2.0
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.7
django-auth-ldap==4.3.0
django-filter==23.1
django-formtools==2.3
django-tables2==2.6.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
dnspython==2.3.0
et-xmlfile==1.1.0
filelock==3.10.7
fonttools==4.47.2
gunicorn==20.1.0
httmock==1.4.0
idna==3.4
importlib-metadata==6.1.0
importlib-resources==6.1.1
jenkinsapi==0.3.13
jira==3.6.0
kiwisolver==1.4.5
Markdown==3.3.4
martor==1.6.26
matplotlib==3.8.2
mysqlclient==2.2.0
numpy==1.26.3
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
pillow==10.2.0
platformdirs==3.2.0
prometheus-api-client==0.5.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.8.2
python-ldap==3.4.3
pytz==2023.3
regex==2023.12.25
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
six==1.16.0
sqlparse==0.2.4
typing-extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.15
webencodings==0.5.1
zipp==3.15.0
